Добавлены задачи для самостоятельного изучения.(Написан тест для открытия какого-то приложения. Написан тесты для открытия приложения, и в нем нажатия на кнопку)
