+++ app/src/androidTest/java/com/volokhinaleksey/popularlibrariesandroid/ui/screens/UserFragmentUIAutomatorTest.kt
@@ -1,4 +1,4 @@
-package com.volokhinaleksey.popularlibrariesandroid.ui.activity
+package com.volokhinaleksey.popularlibrariesandroid.ui.screens
 
 import android.content.Context
 import android.content.Intent
@@ -9,7 +9,7 @@ import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.uiautomator.By
 import androidx.test.uiautomator.UiDevice
 import androidx.test.uiautomator.Until
-import com.google.common.truth.Truth
+import com.google.common.truth.Truth.assertThat
 import org.junit.After
 import org.junit.Before
 import org.junit.Test
@@ -17,7 +17,7 @@ import org.junit.runner.RunWith
 
 @RunWith(AndroidJUnit4::class)
 @SdkSuppress(minSdkVersion = 27)
-class UsersFragmentUIAutomatorTest {
+class UserFragmentUIAutomatorTest {
 
     private val uiDevice = UiDevice.getInstance(InstrumentationRegistry.getInstrumentation())
 
@@ -33,28 +33,34 @@ class UsersFragmentUIAutomatorTest {
         val intent = context?.packageManager?.getLaunchIntentForPackage(packageName)
         intent?.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TASK)
         context?.startActivity(intent)
+        uiDevice.wait(Until.hasObject(By.pkg(packageName).depth(0)), TIMEOUT)
         uiDevice.wait(
-            Until.hasObject(By.pkg(packageName).depth(0)),
-            TIMEOUT
-        )
-    }
-
-    @Test
-    fun check_MainScreenStarted_ReturnTrue() {
-        Truth.assertThat(uiDevice.findObject(By.res(packageName, "users_list_container")))
-            .isNotNull()
-    }
-
-    @Test
-    fun check_NavigationToUserDetails_ReturnTrue() {
-        val navigateItemToUserDetailsScreen = uiDevice.wait(
             Until.findObject(By.res(packageName, "user_login")),
             TIMEOUT
+        ).click()
+    }
+
+    @Test
+    fun check_NavigationToReposDetails_ReturnTrue() {
+        val navigateItemToReposDetailsScreen = uiDevice.wait(
+            Until.findObject(By.res(packageName, "repo_name")),
+            TIMEOUT
         )
-        navigateItemToUserDetailsScreen.click()
-        Truth.assertThat(
+        navigateItemToReposDetailsScreen.click()
+        assertThat(
             uiDevice.wait(
-                Until.findObject(By.res(packageName, "user_details_container")),
+                Until.findObject(By.res(packageName, "repo_details_container")),
+                TIMEOUT
+            )
+        ).isNotNull()
+    }
+
+    @Test
+    fun check_ButtonBackClick_ReturnTrue() {
+        uiDevice.pressBack()
+        assertThat(
+            uiDevice.wait(
+                Until.findObject(By.res(packageName, "users_list_container")),
                 TIMEOUT
             )
         ).isNotNull()
@@ -68,5 +74,4 @@ class UsersFragmentUIAutomatorTest {
     companion object {
         private const val TIMEOUT = 5000L
     }
-
 }
