+++ app/src/androidTest/java/com/volokhinaleksey/popularlibrariesandroid/ui/screens/UsersFragmentUIAutomatorTest.kt
@@ -1,4 +1,4 @@
-package com.volokhinaleksey.popularlibrariesandroid.ui.activity
+package com.volokhinaleksey.popularlibrariesandroid.ui.screens
 
 import android.content.Context
 import android.content.Intent
@@ -9,7 +9,7 @@ import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.uiautomator.By
 import androidx.test.uiautomator.UiDevice
 import androidx.test.uiautomator.Until
-import com.google.common.truth.Truth.assertThat
+import com.google.common.truth.Truth
 import org.junit.After
 import org.junit.Before
 import org.junit.Test
@@ -17,7 +17,7 @@ import org.junit.runner.RunWith
 
 @RunWith(AndroidJUnit4::class)
 @SdkSuppress(minSdkVersion = 27)
-class ReposDetailsFragmentUIAutomatorTest {
+class UsersFragmentUIAutomatorTest {
 
     private val uiDevice = UiDevice.getInstance(InstrumentationRegistry.getInstrumentation())
 
@@ -34,29 +34,25 @@ class ReposDetailsFragmentUIAutomatorTest {
         intent?.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TASK)
         context?.startActivity(intent)
         uiDevice.wait(
+            Until.hasObject(By.pkg(packageName).depth(0)),
+            TIMEOUT
+        )
+    }
+
+    @Test
+    fun check_MainScreenStarted_ReturnTrue() {
+        Truth.assertThat(uiDevice.findObject(By.res(packageName, "users_list_container")))
+            .isNotNull()
+    }
+
+    @Test
+    fun check_NavigationToUserDetails_ReturnTrue() {
+        val navigateItemToUserDetailsScreen = uiDevice.wait(
             Until.findObject(By.res(packageName, "user_login")),
             TIMEOUT
-        ).click()
-        uiDevice.wait(
-            Until.findObject(By.res(packageName, "repo_name")),
-            TIMEOUT
-        ).click()
-    }
-
-    @Test
-    fun check_ButtonOpenInBrowser_ReturnTrue() {
-        val openInBrowser = uiDevice.wait(
-            Until.findObject(By.res(packageName, "open_in_browser")), TIMEOUT
         )
-        openInBrowser.click()
-        uiDevice.waitForWindowUpdate(packageName, 1000)
-        assertThat(uiDevice.currentPackageName).isEqualTo("com.android.chrome")
-    }
-
-    @Test
-    fun check_ButtonBackClick_ReturnTrue() {
-        uiDevice.pressBack()
-        assertThat(
+        navigateItemToUserDetailsScreen.click()
+        Truth.assertThat(
             uiDevice.wait(
                 Until.findObject(By.res(packageName, "user_details_container")),
                 TIMEOUT
